+++ R/spatial_tools.R
@@ -13,6 +13,7 @@
 
 discpoly <- function (center, radius, npoly = 64,
                       class = c("Polygon", "owin", "gpc.poly"),
+                      ## FIXME: should add polygonal "sfg" (or "sfc"?)
                       hole = FALSE)
 {
     class <- match.arg(class)
@@ -35,10 +36,8 @@ discpoly <- function (center, radius, npoly = 64,
     switch(class,
         "Polygon" = Polygon(cbind(c(x,x[1]),c(y,y[1])), hole=hole),
         "gpc.poly" = {
-            ##gpcWarning()
             pts <- list(list(x=x, y=y, hole=hole))
-            if (isClass("gpc.poly") ||
-                (gpclibCheck(FALSE) && requireNamespace("gpclib"))) {
+            if (isClass("gpc.poly")) { #|| requireNamespace("gpclib")
                 new("gpc.poly", pts = pts)
             } else {
                 warning("formal class \"gpc.poly\" not available")
@@ -49,19 +48,23 @@ discpoly <- function (center, radius, npoly = 64,
 }
 
 
-### Wrapper for polyclip or rgeos::gUnaryUnion
+### Wrapper for polyclip or sf::st_union
 
 unionSpatialPolygons <- function (SpP,
-                                  method = c("rgeos", "polyclip"),
+                                  method = c("sf", "polyclip"),
                                   ...)
 {
-    if (identical(method, "gpclib")) {
-        .Deprecated(msg = "method = \"gpclib\" is deprecated; using default")
+    if (identical(method, "gpclib") || identical(method, "rgeos")) {
+        .Deprecated(msg = sprintf("method = \"%s\" is retired; using default", method))
         method <- NULL
     }
     method <- match.arg(method)
     W <- switch(
         method,
+        "sf" = {
+            sf::as_Spatial(sf::st_union(sf::st_as_sfc(SpP), ...),
+                           IDs = "1") # as in rgeos::gUnaryUnion() before
+        },
         "polyclip" = {
             tiles_xylist <- xylist(SpP, reverse=FALSE)
             W_xylist <- polyclip::polyclip(tiles_xylist, tiles_xylist, "union",
@@ -73,14 +76,7 @@ unionSpatialPolygons <- function (SpP,
                        Polygon(cbind(p$x,p$y)[c(1L,length(p$x):1L),])),
                 ID="1")
             SpatialPolygons(list(W_Polygons))
-        },
-        "rgeos" = rgeos::gUnaryUnion(SpP, ...)
-        ## "gpclib" = {
-        ##     gpclibCheck() && maptools::gpclibPermit()
-        ##     maptools::unionSpatialPolygons(
-        ##         SpP, IDs = rep.int(1,length(SpP@polygons)),
-        ##         avoidGEOS = TRUE, ...)
-        ## }
+        }
     )
     ## ensure that W has exactly the same proj4string as SpP
     W@proj4string <- SpP@proj4string
@@ -129,7 +125,7 @@ multiplicity.Spatial <- function (x) multiplicity(coordinates(x))
 
 polyAtBorder <- function (SpP,
                           snap = sqrt(.Machine$double.eps),
-                          method = "rgeos", ...)
+                          method = "sf", ...)
 {
     SpP <- as(SpP, "SpatialPolygons")
     W <- unionSpatialPolygons(SpP, method = method, ...)